Payment and mail services implementation version updates - spring, postgres, keycloak Contact form Added email field to user model UI fixes

+++ catalog-service/src/main/java/com/ElectroMarket/catalogservice/models/Product.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.validation.constraints.*;
 import org.springframework.data.annotation.*;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record Product(
@@ -16,7 +17,7 @@ public record Product(
 
         @NotNull(message = "The product price is required.")
         @Positive(message = "The product price must be greater than zero.")
-        Double price,
+        BigDecimal price,
 
         @NotNull(message = "The product category id is required.")
         Long categoryId,
@@ -55,7 +56,7 @@ public record Product(
         int version
 ) {
     public static Product of(
-            String name, Double price,
+            String name, BigDecimal price,
             Long categoryId, Integer stock,
             String imageUrl, String brand
