+++ order-service/src/main/resources/db/migration/V1__Initial_schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS orders (
     id BIGSERIAL PRIMARY KEY NOT NULL,
-    total float8 NOT NULL,
+    total NUMERIC(10, 2) NOT NULL,
     status VARCHAR(255) NOT NULL,
     created_date TIMESTAMP NOT NULL,
     last_modified_date TIMESTAMP NOT NULL,
